+++ kryten_llm/components/memory/__init__.py
@@ -9,8 +9,9 @@ from kryten_llm.components.memory.embedder import (
     OnnxEmbedder,
     OpenAICompatibleEmbedder,
 )
-from kryten_llm.components.memory.extractor import Fact, FactExtractor
+from kryten_llm.components.memory.extractor import ExtractedFact, Fact, FactExtractor
 from kryten_llm.components.memory.heuristic_extractor import HeuristicFactExtractor
+from kryten_llm.components.memory.llm_extractor import LLMFactExtractor
 from kryten_llm.components.memory.safety import is_safe_message
 from kryten_llm.components.memory.vector_store import (
     VECTOR_STORE_REGISTRY,
@@ -24,8 +25,10 @@ __all__ = [
     "EMBEDDER_REGISTRY",
     "VectorStore",
     "VECTOR_STORE_REGISTRY",
+    "ExtractedFact",
     "Fact",
     "FactExtractor",
     "HeuristicFactExtractor",
+    "LLMFactExtractor",
     "is_safe_message",
 ]